requirements.txt
streamlit
deepgram-sdk
python-dotenv
audio-recorder-streamlit
moviepy
fastapi
uvicorn[standard]
python-multipart
